merge

+++ src/twitchplays/countdown.py
@@ -4,8 +4,8 @@ import time
 class Countdown:
 
     def __init__(self) -> None:
-        self._start_time: int = -1
-        self._countdown_duration: int = 1
+        self._start_time: float = -1
+        self._countdown_duration: float = 1
         self._countdown_active: bool = False
 
     def get_start_time(self) -> float:
@@ -16,8 +16,8 @@ class Countdown:
         self._start_time = time.time()
 
     def get_seconds_remaining(self) -> int:
-        elapsed_time: int = int(time.time() - self._start_time)
-        return max(0, self._countdown_duration - elapsed_time)
+        elapsed_time: float = float(time.time() - self._start_time)
+        return max(0, int(self._countdown_duration - elapsed_time))
 
     def is_countdown_running(self) -> bool:
         return self._countdown_active
